pyproject.toml
[project]
name = "agentic-ai-first"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain[anthropic,google-genai,google-vertexai]>=0.3.25",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
